Promote action plan feature Remove feature flag and deprecated routes, content, locales, specs etc. for the previous user journey from Phase 3. This has all been replaced by the new Phase 4 user journey with the action plan so can be cleaned up.
