+++ app/views/courses/index.html.erb
@@ -1,23 +1,13 @@
 <% page_title :courses_index %>
 
 <% content_for :breadcrumb do
-    if Flipflop.action_plan?
-      generate_breadcrumbs(
-        t('breadcrumb.training_courses_near_you'),
-        [
-          [t('breadcrumb.task_list_home'), task_list_path],
-          [t('breadcrumb.action_plan'), action_plan_path]
-        ]
-      )
-    else
-      generate_breadcrumbs(
-        t('breadcrumb.training_courses_near_you'),
-        [
-          [t('breadcrumb.task_list_home'), task_list_path],
-          [t('breadcrumb.training_hub'), training_hub_path]
-        ]
-      )
-    end
+    generate_breadcrumbs(
+      t('breadcrumb.training_courses_near_you'),
+      [
+        [t('breadcrumb.task_list_home'), task_list_path],
+        [t('breadcrumb.action_plan'), action_plan_path]
+      ]
+    )
   end
 %>
 
@@ -62,7 +52,6 @@
   <div class="govuk-grid-column-one-third">
     <%= render 'shared/contact_us' %>
     <%= render 'shared/save_or_return_to_your_results' %>
-    <%= render 'shared/what_you_can_do_next' unless Flipflop.action_plan? %>
     <%= render 'shared/help_improve_this_service' %>
   </div>
 </div>
