+++ app/views/errors/course_postcode_search_error.html.erb
@@ -5,7 +5,8 @@
     <div class="govuk-grid-column-two-thirds">
       <h1 class="govuk-heading-xl">Sorry, there is a problem with this service</h1>
       <p class="govuk-body">Postcode search isn't working right now.</p>
-      <p class="govuk-body-m">You can still use this service to <%= link_to 'check your existing skills', check_your_skills_path, class: 'govuk-link' %> and get further <%= link_to 'help to find work', next_steps_path, class: 'govuk-link' %>.</p>
+      <!-- TODO: update copy to avoid linking to deprecated route -->
+      <p class="govuk-body-m">You can still use this service to <%= link_to 'check your existing skills', check_your_skills_path, class: 'govuk-link' %> and get further <%#= link_to 'help to find work', next_steps_path, class: 'govuk-link' %>.</p>
       <%= link_to 'Continue', task_list_path, class: 'govuk-button', data: { module: 'govuk-button' } %>
     </div>
   </div>
