--- app/views/job_profiles/_further_help.html.erb
@@ -1,9 +1,0 @@
-<h2 class="govuk-heading-m">Further help to change jobs</h2>
-<p class="govuk-body">You can get help with:</p>
-<ul class="govuk-list govuk-list--bullet">
-  <li>finding work</li>
-  <li>applying for jobs</li>
-  <li>training to prepare you for a job you want to do</li>
-  <li>getting on-the-job training to earn as you learn</li>
-</ul>
-<%= link_to 'Get help changing jobs', next_steps_path, class: 'govuk-button govuk-!-margin-top-1 govuk-!-margin-bottom-6', data: { module: 'govuk-button' } %>
