+++ app/views/job_profiles/_relevant_sections.html.erb
@@ -1,7 +1,6 @@
 <%= render 'skills_section' %>
 <%= @job_profile.section(:day_to_day_tasks).html_safe %>
 <%= @job_profile.section(:working_environment).html_safe %>
-<%= render 'valuable_skills' unless Flipflop.action_plan? %>
 <%= @job_profile.section(:direct_application).html_safe %>
 <%= render 'apprenticeship' %>
 <%= @job_profile.section(:career_path).html_safe %>
@@ -11,10 +10,6 @@
 <%= @job_profile.section(:restrictions_and_requirements).html_safe %>
 <%= render 'jobs_in_your_area' %>
 
-<% if Flipflop.action_plan? %>
-  <h2 class="govuk-heading-m">Select type of work</h2>
-  <p class="govuk-body">Select this as your target type of work and we’ll help you prepare to apply.</p>
-  <%= button_to('Target this type of work', target_job_profile_path(@job_profile.slug), class: 'govuk-button govuk-!-margin-top-1 govuk-!-margin-bottom-6', data: { module: 'govuk-button' }) %>
-<% else %>
-  <%= render 'further_help' %>
-<% end %>
+<h2 class="govuk-heading-m">Select type of work</h2>
+<p class="govuk-body">Select this as your target type of work and we’ll help you prepare to apply.</p>
+<%= button_to('Target this type of work', target_job_profile_path(@job_profile.slug), class: 'govuk-button govuk-!-margin-top-1 govuk-!-margin-bottom-6', data: { module: 'govuk-button' }) %>
