--- app/views/job_profiles/_valuable_skills.html.erb
@@ -1,4 +1,0 @@
-<h2 class="govuk-heading-m">Valuable skills</h2>
-<p class="govuk-body-m">Employers tend to favour candidates with proven maths and English skills. Completing a short course in maths and English can really boost your employability.</p>
-<%= link_to('Find a training course', training_hub_path, class: 'govuk-button govuk-!-margin-top-1  govuk-!-margin-bottom-6', data: { cy: 'find-a-course-btn', module: 'govuk-button' }) %>
-<hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
