--- app/views/pages/english_overview.html.erb
@@ -1,32 +1,0 @@
-<% page_title :pages_english_overview %>
-<% content_for :breadcrumb do
-    generate_breadcrumbs(
-      t('breadcrumb.english_overview'),
-      [
-        [t('breadcrumb.task_list_home'), task_list_path],
-        [t('breadcrumb.training_hub'), training_hub_path]
-      ]
-    )
-  end
-%>
-
-<div class="govuk-grid-row govuk-!-margin-top-7 govuk-body">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl"><%= t('.title') %></h1>
-    <h2 class="govuk-heading-m">How will an English course help my career?</h2>
-    <p class="govuk-body">Employers want to know that you can communicate confidently and clearly in the workplace. Completing an English course helps show them you have the skills they want. </p>
-    <p class="govuk-body">English classes will help you develop skills in things such as spelling and grammar, reading and writing, as well as speaking and listening skills. They’ll help improve your confidence and show that you can apply those skills in the workplace.</p>
-    <p class="govuk-body">Studying at adult level isn’t like being at school. You’ll find the course focuses on giving you the practical skills you need to succeed at work.</p>
-    <h2 class="govuk-heading-m">The right course for you</h2>
-    <p class="govuk-body">There are lots of English courses available for every level of ability. The course provider will help assess what level of course is right for you. You’ll also be able to choose a course that offers flexible learning or part time study to fit around your life.</p>
-    <p class="govuk-body">Many courses are available for free in your area.</p>
-    <%= link_to 'Find an English course', courses_path('english'), class: 'govuk-button govuk-!-margin-top-5', data: { module: 'govuk-button' } %>
-  </div>
-
-  <div class="govuk-grid-column-one-third">
-    <%= render "shared/contact_us" %>
-    <%= render 'shared/save_or_return_to_your_results' %>
-    <%= render 'shared/what_you_can_do_next' %>
-    <%= render 'shared/help_improve_this_service' %>
-  </div>
-</div>
