--- app/views/pages/next_steps.html.erb
@@ -1,53 +1,0 @@
-<% page_title :pages_next_steps %>
-<% content_for :breadcrumb do
-    generate_breadcrumbs(
-      t('breadcrumb.next_steps'),
-      [
-        [t('breadcrumb.task_list_home'), task_list_path]
-      ]
-    )
-  end
-%>
-
-<div class="govuk-grid-row govuk-!-margin-top-7 govuk-body">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl"><%= t('pages.next_steps.title') %></h1>
-    <h3 class="govuk-heading-m govuk-!-margin-bottom-1">Find advertised jobs</h3>
-    <p class="govuk-body">Find what jobs are available in your area using the <%= link_to 'Find a job', 'https://findajob.dwp.gov.uk/', class: 'govuk-link', data: { tracked_event: 'Next steps - Clicked Find a job', tracked_event_props: { url: 'https://findajob.dwp.gov.uk/' } }, target: '_blank' %> service from the Department for Work and Pensions.</p>
-    <p class="govuk-body">The <%= link_to 'National Careers Service', 'https://nationalcareers.service.gov.uk/get-a-job', class: 'govuk-link', data: { tracked_event: 'Next steps - Clicked National Careers Service', tracked_event_props: { url: 'https://nationalcareers.service.gov.uk/get-a-job' } }, target: '_blank' %> also has information on where and how to find work.</p>
-
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <h3 class="govuk-heading-m govuk-!-margin-bottom-1">Prepare to apply</h3>
-    <p class="govuk-body">Learn how to <%= link_to 'write an effective CV', 'https://nationalcareers.service.gov.uk/get-a-job/cv-sections', class: 'govuk-link', data: { tracked_event: 'Next steps - write an effective CV', tracked_event_props: { url: 'https://nationalcareers.service.gov.uk/get-a-job/cv-sections' } }, target: '_blank' %>.</p>
-    <p class="govuk-body">Learn how to <%= link_to 'write a covering letter', 'https://nationalcareers.service.gov.uk/get-a-job/covering-letter', class: 'govuk-link', data: { tracked_event: 'Next steps - write a covering letter', tracked_event_props: { url: 'https://nationalcareers.service.gov.uk/get-a-job/covering-letter' } }, target: '_blank' %>.</p>
-    <p class="govuk-body">Find out how to <%= link_to 'prepare for interviews', 'https://nationalcareers.service.gov.uk/get-a-job/interview-advice', class: 'govuk-link', data: { tracked_event: 'Next steps - prepare for interviews', tracked_event_props: { url: 'https://nationalcareers.service.gov.uk/get-a-job/interview-advice' } }, target: '_blank' %>.</p>
-
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <h3 class="govuk-heading-m govuk-!-margin-bottom-1">Find other training</h3>
-    <p class="govuk-body">Training can help you access a broader range of jobs and stay in secure work. Use the National Careers Service site to find <%= link_to 'training', 'https://nationalcareers.service.gov.uk/find-a-course', class: 'govuk-link', data: { tracked_event: 'Next steps - training', tracked_event_props: { url: 'https://nationalcareers.service.gov.uk/find-a-course' } }, target: '_blank' %> in your area for specific jobs and to raise your general employability.</p>
-
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <h3 class="govuk-heading-m govuk-!-margin-bottom-1">On-the-job training</h3>
-    <p class="govuk-body">If you have existing experience and want to stay in work as you learn, it's a great idea to train on-the-job in a new type of role. There's no age limit for these opportunities and you can often work and learn at your existing level of seniority. <%= link_to 'Find an opportunity in your area', 'https://www.gov.uk/apprenticeships-guide', class: 'govuk-link', data: { tracked_event: 'Next steps - Find an opportunity in your area', tracked_event_props: { url: 'https://www.gov.uk/apprenticeships-guide' } }, target: '_blank' %></p>
-
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-
-    <h3 class="govuk-heading-m govuk-!-margin-bottom-4">On offer near you</h3>
-    <p class="govuk-heading-s govuk-!-margin-bottom-1">Cambridge & Peterborough</p>
-    <p class="govuk-body">The <%= link_to 'Health and Care Sector Work Academy', 'https://cambridgeshirepeterborough-ca.gov.uk/about-us/programmes/apprenticeships-and-skills/health-and-care-sector-work-academy/', class: 'govuk-link', data: { tracked_event: 'Next steps - Health and Care Sector Work Academy', tracked_event_props: { url: 'https://cambridgeshirepeterborough-ca.gov.uk/about-us/programmes/apprenticeships-and-skills/health-and-care-sector-work-academy/' } }, target: '_blank' %> helps you move into the health and care sector or progress further if you're already working in this area.</p>
-    <p class="govuk-heading-s govuk-!-margin-bottom-1">Leeds</p>
-    <p class="govuk-body">Information on changing careers from <%= link_to 'Leeds City Region Local Enterprise Partnership', 'https://futuregoals.co.uk/', class: 'govuk-link',  data: { tracked_event: 'Next steps - Leeds City Region Local Enterprise Partnership', tracked_event_props: { url: 'https://futuregoals.co.uk/' } }, target: '_blank' %>.</p>
-
-    <p class="govuk-heading-s govuk-!-margin-top-4 govuk-!-margin-bottom-1">Liverpool</p>
-    <p class="govuk-body">Use the <%= link_to 'Be More', 'https://be-more.info', class:'govuk-link', data: { tracked_event: 'Next steps - Be More', tracked_event_props: { url: 'https://be-more.info' } }, target: '_blank' %> service to find apprenticeships in the Liverpool area.</p>
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-  </div>
-  <div class="govuk-grid-column-one-third">
-    <%= render "shared/contact_us" %>
-    <%= render 'shared/save_or_return_to_your_results' %>
-    <%= render "shared/help_improve_this_service" %>
-  </div>
-</div>
