--- app/views/pages/task_list_items/v1/_item_3.html.erb
@@ -1,15 +1,0 @@
-<li>
-  <h2 class="app-task-list__section">
-    <span class="app-task-list__section-number">3. </span> Find training that's right for you
-  </h2>
-  <ul class="app-task-list__items">
-    <li class="app-task-list__item">
-      <div class="app-task-list__task-name">
-        <%= link_or_blocked(path: training_hub_path, enabled: user_session.job_profile_skills?, span_id: 3) do %>
-          Find a training course
-        <% end %>
-        <p class="govuk-body govuk-!-margin-bottom-0">Training can help you access a broader range of jobs and stay in secure work.</p>
-      </div>
-    </li>
-  </ul>
-</li>
