--- app/views/pages/task_list_items/v1/_item_4.html.erb
@@ -1,15 +1,0 @@
-<li>
-  <h2 class="app-task-list__section">
-    <span class="app-task-list__section-number">4. </span> Get help changing jobs
-  </h2>
-  <ul class="app-task-list__items">
-    <li class="app-task-list__item">
-      <div class="app-task-list__task-name">
-        <%= link_or_blocked(path: next_steps_path, enabled: user_session.job_profile_skills?, span_id: 4) do %>
-          Further help to find work
-        <% end %>
-        <p class="govuk-body govuk-!-margin-bottom-0">Get help writing a CV and cover letter and preparing for interview. See if on-the-job training is right for you and find local support.</p>
-      </div>
-    </li>
-  </ul>
-</li>
