--- app/views/pages/task_list_items/v1/_item_5.html.erb
@@ -1,15 +1,0 @@
-<li>
-  <h2 class="app-task-list__section">
-    <span class="app-task-list__section-number">5. </span> Help improve this service
-  </h2>
-  <ul class="app-task-list__items">
-    <li class="app-task-list__item">
-      <div class="app-task-list__task-name">
-        <%= link_or_blocked(path: 'https://www.smartsurvey.co.uk/s/get-help-to-retrain/', enabled: user_session.job_profile_skills?, span_id: 5, data: { tracked_event: 'TaskList - User Exit Survey Link' }, target: '_blank') do %>
-          Take a quick survey
-        <% end %>
-        <p class="govuk-body govuk-!-margin-bottom-0">Tell us about your experience using Get help to retrain.</p>
-      </div>
-    </li>
-  </ul>
-</li>
