--- app/views/pages/training_hub.html.erb
@@ -1,35 +1,0 @@
-<% page_title :pages_training_hub %>
-<% content_for :breadcrumb do
-    generate_breadcrumbs(
-      t('breadcrumb.training_hub'),
-      [
-        [t('breadcrumb.task_list_home'), task_list_path]
-      ]
-    )
-  end
-%>
-
-<div class="govuk-grid-row govuk-!-margin-top-7 govuk-body">
-  <div class="govuk-grid-column-two-thirds govuk-!-margin-bottom-7">
-    <h1 class="govuk-heading-xl">Find training that boosts your job options</h1>
-    <p class="govuk-body-l">You could get a better job by adding to the skills you already have. Complete a short course in maths or English and you could find you have a wider range of job options. Many courses are free and offer flexible times to fit round your life.</p>
-    <p class="govuk-body">This is a new scheme - a wider range of courses will be added in future.</p>
-    <p class="govuk-body">You can also find on-the-job training that lets you earn while you learn.</p>
-    <h2 class="govuk-heading-l">What’s on offer?</h2>
-    <%= render 'shared/training_hub/maths' %>
-    <%= render 'shared/training_hub/english' %>
-    <h2 class="govuk-heading-m govuk-!-margin-bottom-1">On-the-job training</h2>
-    <p class="govuk-body-m govuk-!-margin-bottom-1">Gain skills and experience while continuing to earn. There's no age limit and different levels of study are available. This means on-the-job-training is suitable for experienced older workers wanting to stay in work as they learn.</p>
-    <%= link_to('Find an opportunity', 'https://www.gov.uk/apprenticeships-guide', class: 'govuk-button govuk-!-margin-top-5 govuk-button govuk-!-margin-bottom-2', target: '_blank', data: { tracked_event: "Training hub - Clicked 'Find an opportunity': https://www.gov.uk/apprenticeships-guide", module: 'govuk-button' }) %>
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-    <h2 class="govuk-heading-m govuk-!-margin-bottom-1">More help to change jobs</h2>
-    <p class="govuk-body-m govuk-!-margin-bottom-1">Get support to find and apply for jobs, or find other ways to move into a new role.</p>
-    <%= link_to('Other ways to change jobs', next_steps_path, class: 'govuk-button govuk-!-margin-top-5 govuk-button govuk-!-margin-bottom-2', data: { module: 'govuk-button' }) %>
-    <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
-  </div>
-  <div class="govuk-grid-column-one-third">
-    <%= render "shared/contact_us" %>
-    <%= render 'shared/save_or_return_to_your_results' %>
-    <%= render 'shared/help_improve_this_service' %>
-  </div>
-</div>
