--- app/views/shared/_what_you_can_do_next.html.erb
@@ -1,6 +1,0 @@
-<div class="govuk-related-items govuk-!-padding-bottom-4 top-border__container">
-  <h2 class="govuk-heading-m"><%= t('what_you_can_do_next.title') %></h2>
-  <p class="govuk-body-m">
-    <%= link_to t('what_you_can_do_next.cta_copy'), next_steps_path, class: 'govuk-link' %>
-  </p>
-</div>
