--- app/views/shared/training_hub/_english.html.erb
@@ -1,6 +1,0 @@
-<h2 class="govuk-heading-m govuk-!-margin-bottom-1">English</h2>
-<p class="govuk-body-m govuk-!-margin-bottom-1">Showing that you can communicate well at work is really important for your job prospects. English courses focus on giving you the skills you need to succeed in the workplace.</p>
-<%= link_to('More information on English courses', english_course_overview_path, class: 'govuk-body-m govuk-link') %>
-<br>
-<%= link_to('Find an English course', courses_path('english'), class: 'govuk-button govuk-!-margin-top-5 govuk-button govuk-!-margin-bottom-2', data: { cy: 'find-a-english-course-btn', module: 'govuk-button' }) %>
-<hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
