--- app/views/shared/training_hub/_maths.html.erb
@@ -1,6 +1,0 @@
-<h2 class="govuk-heading-m govuk-!-margin-bottom-1">Maths</h2>
-<p class="govuk-body-m govuk-!-margin-bottom-1">Employers tell us they are looking for people who can show they have some maths skills. Completing a free maths course can really boost your job prospects in many different types of work. </p>
-<%= link_to('More information on maths courses', maths_course_overview_path, class: 'govuk-body-m govuk-link') %>
-<br>
-<%= link_to('Find a maths course', courses_path('maths'), class: 'govuk-button govuk-!-margin-top-5 govuk-button govuk-!-margin-bottom-2', data: { cy: 'find-a-maths-course-btn', module: 'govuk-button' }) %>
-<hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
