+++ app/views/skills_matcher/_no_results.html.erb
@@ -1,3 +1,4 @@
 <h1 class="govuk-heading-xl">More information needed</h1>
 <p class="govuk-body">You'll get more job suggestions if you <%= link_to 'add more skills from another role', check_your_skills_path, class: 'govuk-link' %> you’ve done.</p>
-<p class="govuk-body">If you haven't done any other jobs, doing some training will increase your job options. <%= link_to 'Find a training course', training_hub_path, class: 'govuk-link' %></p>
+<!-- TODO: update copy to avoid linking to deprecated route -->
+<p class="govuk-body">If you haven't done any other jobs, doing some training will increase your job options. <%#= link_to 'Find a training course', training_hub_path, class: 'govuk-link' %></p>
