+++ config/features.rb
@@ -5,5 +5,4 @@ Flipflop.configure do
 
   feature :foo, description: 'Example feature flag', default: true
   feature :health_check, description: 'Dummy feature for split.io health check'
-  feature :action_plan, description: 'Action plan and updated job profile page'
 end
