+++ config/locales/en.yml
@@ -10,14 +10,10 @@ en-GB:
     task_list_home: 'Home: Get help to retrain'
     check_your_skills: Check your existing skills
     job_matches: Your job matches
-    next_steps: Further help
     search_results: Search results
     job_profile: Job profile details
     your_skills: Your skills
     training_courses_near_you: Find and apply to training courses near you
-    maths_overview: Benefits of maths
-    english_overview: Benefits of English
-    training_hub: Find training
     location_ineligible: No courses in your area
     current_job_skills: Job skills
     user_personal_data: Your information
@@ -42,10 +38,6 @@ en-GB:
     job_profiles_show: Get help to retrain - Job profile
     courses_index: Get help to retrain - Courses near me
     pages_task_list: Get help to retrain - Tasks
-    pages_next_steps: Get help to retrain - Further help
-    pages_maths_overview: Get help to retrain - Benefits of maths
-    pages_english_overview: Get help to retrain - Benefits of English
-    pages_training_hub: Get help to retrain - Training Hub
     pages_cookies_policy: Get help to retrain - Cookies policy
     pages_privacy_policy: Get help to retrain - Privacy policy
     pages_location_ineligible: Get help to retrain - No courses in your area
@@ -123,22 +115,12 @@ en-GB:
       title: Find local schemes and offers
     task_list:
       title: Step-by-step guide to changing types of work
-    next_steps:
-      title: Get help changing jobs
-    maths_overview:
-      title: Benefits of doing a maths course
-    english_overview:
-      title: Benefits of doing an English course
     location_ineligible:
       title: The full Get help to retrain service isn't available in your area yet
     cookies_policy:
       title: Cookies
     privacy_policy:
       title: Privacy policy
-  what_you_can_do_next:
-    title: Further help to find work
-    sub_title: Support to move into a new role.
-    cta_copy: Get help changing jobs
   help_improve_this_service:
     title: Help improve this service
     body: Take a quick survey and tell us about your experience
