+++ spec/features/job_profile_spec.rb
@@ -1,10 +1,6 @@
 require 'rails_helper'
 
 RSpec.feature 'Job profile spec' do
-  background do
-    disable_feature! :action_plan
-  end
-
   scenario 'Page has correct title including job profile name' do
     create(:job_profile, :with_html_content, name: 'Acrobat', slug: 'acrobat')
 
@@ -21,56 +17,12 @@ RSpec.feature 'Job profile spec' do
     expect(page).to have_content('Other similar jobs include: therapy master, dog walker, trainer')
   end
 
-  scenario 'Page contains the Valuable skills section' do
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    expect(page).to have_content('Valuable skills')
-  end
-
-  scenario 'User navigates to training hub page from valuable skills section' do
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    click_on('Find a training course')
-
-    expect(page).to have_current_path(training_hub_path)
-  end
-
-  scenario 'Page contains the Further help to change jobs section' do
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    expect(page).to have_content('Further help to change jobs')
-  end
-
-  scenario 'User navigates to next steps page from valuable skills section' do
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    click_on('Get help changing jobs')
-
-    expect(page).to have_current_path(next_steps_path)
-  end
-
   scenario 'Apprenticeship section is not rendered unless it exists' do
     job_profile = create(:job_profile, :with_html_content)
 
     visit(job_profile_path(job_profile.slug))
 
     expect(page).not_to have_content('Apprenticeship')
-  end
-
-  scenario 'Page does not have CTA: Target this type of work' do
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    expect(page).not_to have_button('Target this type of work')
   end
 
   scenario 'User can see number of job vacancies near them' do
@@ -131,29 +83,7 @@ RSpec.feature 'Job profile spec' do
     expect(page).not_to have_content('Jobs in your area')
   end
 
-  scenario 'User no longer sees the Valuable skills section on V2' do
-    enable_feature! :action_plan
-
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    expect(page).not_to have_content('Valuable skills')
-  end
-
-  scenario 'User no longer sees the Further help to change jobs section on V2' do
-    enable_feature! :action_plan
-
-    job_profile = create(:job_profile, :with_html_content)
-
-    visit(job_profile_path(job_profile.slug))
-
-    expect(page).not_to have_content('Further help to change jobs')
-  end
-
-  scenario 'User can target a job profile on V2' do
-    enable_feature! :action_plan
-
+  scenario 'User can target a job profile' do
     job_profile = create(:job_profile, :with_html_content)
 
     visit(job_profile_path(job_profile.slug))
