+++ spec/features/user_registration_spec.rb
@@ -110,14 +110,14 @@ RSpec.feature 'User registration' do
     visit(job_profile_skills_path(job_profile_id: job_profile.slug))
     click_on('Select these skills')
     click_on('Save my results')
-    visit(next_steps_path)
+    visit(skills_matcher_index_path)
     click_on('Save my results')
     fill_in('email', with: 'test@test.test')
     click_on('Save your results')
     click_on('send it again')
     click_on('Continue')
 
-    expect(page).to have_current_path(next_steps_path)
+    expect(page).to have_current_path(skills_matcher_index_path)
   end
 
   scenario 'When NotifyService is down, the user sees the correct error page when one tries to save results' do
@@ -137,27 +137,27 @@ RSpec.feature 'User registration' do
 
     visit(job_profile_skills_path(job_profile_id: job_profile.slug))
     click_on('Select these skills')
-    visit(training_hub_path)
+    visit(skills_matcher_index_path)
     click_on('Save my results')
     fill_in('email', with: 'test@test.test')
     click_on('Save your results')
 
     click_on('Continue')
 
-    expect(page).to have_current_path(training_hub_path)
+    expect(page).to have_current_path(skills_matcher_index_path)
   end
 
   scenario 'User can resume their journey to where they first clicked save my results' do
     visit(job_profile_skills_path(job_profile_id: job_profile.slug))
     click_on('Select these skills')
     click_on('Save my results')
-    visit(next_steps_path)
+    visit(skills_matcher_index_path)
     click_on('Save my results')
     fill_in('email', with: 'test@test.test')
     click_on('Save your results')
     click_on('Continue')
 
-    expect(page).to have_current_path(next_steps_path)
+    expect(page).to have_current_path(skills_matcher_index_path)
   end
 
   scenario 'User redirected to task list page if they directly linked to email sent page' do
