+++ spec/routing/job_profiles_spec.rb
@@ -9,23 +9,7 @@ RSpec.describe 'routes for Job Profiles', type: :routing do
     expect(delete('/job-profiles/fishy')).to route_to(controller: 'job_profiles', action: 'destroy', id: 'fishy')
   end
 
-  context 'with action_plan feature enabled' do
-    before do
-      enable_feature! :action_plan
-    end
-
-    it 'successfully routes to job_profiles#target' do
-      expect(post('/job-profiles/fishy/target')).to route_to(controller: 'job_profiles', action: 'target', id: 'fishy')
-    end
-  end
-
-  context 'with action_plan feature disabled' do
-    before do
-      disable_feature! :action_plan
-    end
-
-    it 'does not root to job_profiles#target' do
-      expect(post('/job-profiles/fishy/target')).not_to be_routable
-    end
+  it 'successfully routes to job_profiles#target' do
+    expect(post('/job-profiles/fishy/target')).to route_to(controller: 'job_profiles', action: 'target', id: 'fishy')
   end
 end
