+++ spec/routing/job_vacancies_spec.rb
@@ -1,31 +1,11 @@
 require 'rails_helper'
 
 RSpec.describe 'routes for job vacancies', type: :routing do
-  context 'with action_plan feature enabled' do
-    before do
-      enable_feature! :action_plan
-    end
-
-    it 'successfully routes to get job_vacancies#index' do
-      expect(get('/jobs-near-me')).to route_to(controller: 'job_vacancies', action: 'index')
-    end
-
-    it 'successfully routes to post job_vacancies#index' do
-      expect(post('/jobs-near-me')).to route_to(controller: 'job_vacancies', action: 'index')
-    end
+  it 'successfully routes to get job_vacancies#index' do
+    expect(get('/jobs-near-me')).to route_to(controller: 'job_vacancies', action: 'index')
   end
 
-  context 'with action_plan feature disabled' do
-    before do
-      disable_feature! :action_plan
-    end
-
-    it 'does not root to get job_vacancies#index' do
-      expect(get('/jobs-near-me')).not_to be_routable
-    end
-
-    it 'does not root to post job_vacancies#index' do
-      expect(post('/jobs-near-me')).not_to be_routable
-    end
+  it 'successfully routes to post job_vacancies#index' do
+    expect(post('/jobs-near-me')).to route_to(controller: 'job_vacancies', action: 'index')
   end
 end
